优化MachO

+++ src/main/java/cn/banny/emulator/ios/MachO.java
@@ -2,8 +2,8 @@ package cn.banny.emulator.ios;
 
 interface MachO {
 
-    int CPU_SUBTYPE_ARM_V6 = 6;
+    long CPU_SUBTYPE_ARM_V6 = 6;
 
-    int CPU_SUBTYPE_ARM_V7 = 9;
+    long CPU_SUBTYPE_ARM_V7 = 9;
 
 }
